free air control using opengl
The main aim of this Mini Project is to illustrate the concepts and usage of Air Traffic Control in OpenGL. Air traffic control (ATC) is a service provided by ground-based controllers who direct aircraft on the ground and through controlled airspace. The primary purpose of ATC systems worldwide is to prevent collisions, organize and expedite the flow of traffic, and provide information and other support for pilots To prevent collisions, ATC enforces traffic separation rules, which ensure each aircraft maintains a minimum amount of empty space around it at all times. We have used input devices like the mouse and keyboard to interact with the program. We have used input devices like the mouse and keyboard to interact with the program. We have added menu which makes the program more interactive.